Lesson11 corrected

+++ src/main/java/by/it_academy/lesson11/Sensor.java
@@ -8,6 +8,6 @@ public interface Sensor {
 
     void setOff();
 
-    double read();
+    int read();
 
 }
